fix : 카테고리 CRUD
- CategoryController.java
1. @AuthenticationPrincipal

- CategoryService.java
1. BookResponseDto
형태에 맞게 조회 결과 수정

+++ src/main/java/com/sparta/bookflex/domain/category/controller/CategoryController.java
@@ -1,16 +1,24 @@
 package com.sparta.bookflex.domain.category.controller;
 
 import com.sparta.bookflex.common.aop.Envelop;
+import com.sparta.bookflex.common.exception.BusinessException;
+import com.sparta.bookflex.common.security.UserDetailsImpl;
 import com.sparta.bookflex.domain.category.dto.CategoryRequestDto;
 import com.sparta.bookflex.domain.category.dto.CategoryResponseDto;
 import com.sparta.bookflex.domain.category.service.CategoryService;
+import com.sparta.bookflex.domain.user.entity.User;
+import com.sparta.bookflex.domain.user.enums.UserRole;
+import com.sparta.bookflex.domain.user.service.AuthService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static com.sparta.bookflex.common.exception.ErrorCode.USER_NOT_AUTHENTICATED;
 
 @RestController
 @RequiredArgsConstructor
@@ -18,15 +26,17 @@ import java.util.List;
 public class CategoryController {
 
     private final CategoryService categoryService;
+    private final AuthService authService;
 
     @Envelop("카테고리를 추가했습니다.")
     @PostMapping
-    public ResponseEntity<CategoryResponseDto> createCategory(// @AuthenticationPrincipal UserDetailsImpl userDetails,
+    public ResponseEntity<CategoryResponseDto> createCategory(@AuthenticationPrincipal UserDetailsImpl userDetails,
                                                               @RequestBody @Valid CategoryRequestDto requestDto) {
 
-//        userRepository.findById(userDetails.getUser().getId()).orElseThrow(
-//                () -> new BusinessException(USER_NOT_AUTHENTICATED)
-//        );
+        User user = authService.findByUserName(userDetails.getUser().getUsername());
+        if (user.getAuth() != UserRole.ADMIN) {
+            throw new BusinessException(USER_NOT_AUTHENTICATED);
+        }
 
         CategoryResponseDto responseDto = categoryService.createCategory(requestDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
@@ -35,14 +45,10 @@ public class CategoryController {
 
     @GetMapping("/{categoryId}")
     @Envelop("카테고리 조회에 성공하였습니다.")
-    public ResponseEntity<CategoryResponseDto> getSingleCategory(//@AuthenticationPrincipal UserDetailsImpl userDetails,
+    public ResponseEntity<CategoryResponseDto> getSingleCategory(@AuthenticationPrincipal UserDetailsImpl userDetails,
                                                                  @PathVariable long categoryId,
                                                                  @RequestParam(value = "page", defaultValue = "1") int page,
                                                                  @RequestParam(value = "sortBy", defaultValue = "createdAt") String sortBy) {
-
-//        userRepository.findById(userDetails.getUser().getId()).orElseThrow(
-//                () -> new BusinessException(USER_NOT_AUTHENTICATED)
-//        );
 
         CategoryResponseDto responseDto = categoryService.getSingleCategory(categoryId, page - 1, sortBy);
         return ResponseEntity.status(HttpStatus.OK).body(responseDto);
@@ -50,13 +56,9 @@ public class CategoryController {
 
     @GetMapping
     @Envelop("카테고리 목록 조회에 성공하였습니다.")
-    public ResponseEntity<List<CategoryResponseDto>> getAllCategories(//@AuthenticationPrincipal UserDetailsImpl userDetails,
+    public ResponseEntity<List<CategoryResponseDto>> getAllCategories(@AuthenticationPrincipal UserDetailsImpl userDetails,
                                                                       @RequestParam(value = "page", defaultValue = "1") int page,
                                                                       @RequestParam(value = "sortBy", defaultValue = "createdAt") String sortBy) {
-
-//        userRepository.findById(userDetails.getUser().getId()).orElseThrow(
-//                () -> new BusinessException(USER_NOT_AUTHENTICATED)
-//        );
 
         List<CategoryResponseDto> responseList = categoryService.getAllCategories(page - 1, sortBy);
         return ResponseEntity.status(HttpStatus.OK).body(responseList);
@@ -64,13 +66,14 @@ public class CategoryController {
 
     @PutMapping("/{categoryId}")
     @Envelop("카테고리를 수정했습니다.")
-    public ResponseEntity<CategoryResponseDto> updateCategory(//@AuthenticationPrincipal UserDetailsImpl userDetails,
+    public ResponseEntity<CategoryResponseDto> updateCategory(@AuthenticationPrincipal UserDetailsImpl userDetails,
                                                               @PathVariable long categoryId,
                                                               @RequestBody @Valid CategoryRequestDto requestDto) {
 
-//        userRepository.findById(userDetails.getUser().getId()).orElseThrow(
-//                () -> new BusinessException(USER_NOT_AUTHENTICATED)
-//        );
+        User user = authService.findByUserName(userDetails.getUser().getUsername());
+        if (user.getAuth() != UserRole.ADMIN) {
+            throw new BusinessException(USER_NOT_AUTHENTICATED);
+        }
 
         CategoryResponseDto responseDto = categoryService.updateCategory(categoryId, requestDto);
         return ResponseEntity.status(HttpStatus.OK).body(responseDto);
@@ -79,12 +82,13 @@ public class CategoryController {
 
     @DeleteMapping("/{categoryId}")
     @Envelop("카테고리를 삭제했습니다.")
-    public ResponseEntity deleteCategory(//@AuthenticationPrincipal UserDetailsImpl userDetails,
+    public ResponseEntity deleteCategory(@AuthenticationPrincipal UserDetailsImpl userDetails,
                                          @PathVariable long categoryId) {
 
-//        userRepository.findById(userDetails.getUser().getId()).orElseThrow(
-//                () -> new BusinessException(USER_NOT_AUTHENTICATED)
-//        );
+        User user = authService.findByUserName(userDetails.getUser().getUsername());
+        if (user.getAuth() != UserRole.ADMIN) {
+            throw new BusinessException(USER_NOT_AUTHENTICATED);
+        }
 
         categoryService.deleteCategory(categoryId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
